+++ public/js/jsvalidation.js
@@ -2448,6 +2448,9 @@ laravelValidation = {
             } else {
                 cache[element.name][name]={};
                 var nameParts = name.split("[*]");
+                if (nameParts.length==1) {
+                    nameParts.push('');
+                }
                 var regexpParts = nameParts.map(function(currentValue, index) {
                     if (index % 2 === 0) {
                         currentValue = currentValue + '[';
@@ -2725,16 +2728,27 @@ $.extend(true, laravelValidation, {
             }
 
             var validator = $.data(element.form, "validator");
-            var objRules = validator.settings.rules[element.name];
-            if ('laravelValidation' in objRules) {
-                var _rules=objRules.laravelValidation;
-                for (var i = 0; i < _rules.length; i++) {
-                    if ($.inArray(_rules[i][0],rules) !== -1) {
-                        found = true;
-                        break;
+            var listRules = [];
+            var cache = validator.arrayRulesCache;
+            if (element.name in cache) {
+                $.each(cache[element.name], function (index, arrayRule) {
+                    listRules.push(arrayRule);
+                });
+            }
+            if (element.name in validator.settings.rules) {
+                listRules.push(validator.settings.rules[element.name]);
+            }
+            $.each(listRules, function(index,objRules){
+                if ('laravelValidation' in objRules) {
+                    var _rules=objRules.laravelValidation;
+                    for (var i = 0; i < _rules.length; i++) {
+                        if ($.inArray(_rules[i][0],rules) !== -1) {
+                            found = true;
+                            return false;
+                        }
                     }
                 }
-            }
+            });
 
             return found;
         },
@@ -2766,7 +2780,7 @@ $.extend(true, laravelValidation, {
             } else if ($.isArray(value)) {
                 return parseFloat(value.length);
             } else if (element.type === 'file') {
-                return parseFloat(Math.ceil(this.fileinfo(element).size));
+                return parseFloat(Math.floor(this.fileinfo(element).size));
             }
 
             return parseFloat(this.strlen(value));
@@ -3452,6 +3466,14 @@ $.extend(true, laravelValidation, {
             return true;
         },
 
+        /**
+         * Bail This is the default behaivour os JSValidation.
+         * Always returns true, just lets us put sometimes in rules.*
+         * @return {boolean}
+         */
+        Bail: function() {
+            return true;
+        },
 
         /**
          * Validate the given attribute is filled if it is present.